feat: cr and tdl to separate store files
